add textvqa to evaluation

+++ open_flamingo/eval/eval_datasets.py
@@ -146,7 +146,7 @@ class VQADataset(Dataset):
             image = Image.open(img_path)
             image.load()
         results = {
-            "image": image.half(),
+            "image": image,
             "question": question["question"],
             "question_id": question["question_id"],
         }
